Fixing image cache
