+++ lib/tools/repository/logo_repository.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 
 import 'package:f_logs/f_logs.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:http_parser/http_parser.dart';
 import 'package:myecl/tools/exception.dart';
 import 'package:myecl/tools/repository/repository.dart';
@@ -13,14 +14,14 @@ import 'package:path_provider/path_provider.dart';
 abstract class LogoRepository extends Repository {
   static const String expiredTokenDetail = "Could not validate credentials";
 
-  Future<Image> getLogo(String id, {String suffix = ""}) async {
+  Future<Uint8List> getLogo(String id, {String suffix = ""}) async {
     try {
       final response =
           await http.get(Uri.parse("$host$ext$id$suffix"), headers: headers);
       if (response.statusCode == 200) {
         try {
           await cacheManager.writeImage(ext + id + suffix, response.bodyBytes);
-          return Image.memory(response.bodyBytes);
+          return response.bodyBytes;
         } catch (e) {
           FLog.error(
               text: "GET $ext$id$suffix\nError while decoding response",
@@ -60,9 +61,8 @@ abstract class LogoRepository extends Repository {
     }
   }
 
-  Future<Image> addLogo(String path, String id, {String suffix = ""}) async {
+  Future<Uint8List> addLogo(String path, String id, {String suffix = ""}) async {
     final file = File(path);
-    final image = Image.file(file);
     final request =
         http.MultipartRequest('POST', Uri.parse("$host$ext$id$suffix"))
           ..headers.addAll(headers)
@@ -91,7 +91,7 @@ abstract class LogoRepository extends Repository {
         throw AppException(ErrorType.notFound, value);
       }
     });
-    return image;
+    return file.readAsBytes();
   }
 
   Future<File> saveLogoToTemp(String path) async {
