+++ lib/vote/repositories/pretendance_logo_repository.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:myecl/tools/repository/logo_repository.dart';
 
 class PretendanceLogoRepository extends LogoRepository {
@@ -7,10 +8,15 @@ class PretendanceLogoRepository extends LogoRepository {
   final ext = 'campaign/lists/';
 
   Future<Image> getPretendenceLogo(String id) async {
-    return await getLogo(id, suffix: "/logo");
+    final uint8List =  await getLogo(id, suffix: "/logo");
+    if (uint8List.isEmpty) {
+      return Image.asset("assets/images/logo.png");
+    }
+    return Image.memory(uint8List);
   }
 
   Future<Image> addPretendenceLogo(String path, String id) async {
-    return await addLogo(path, id, suffix: "/logo");
+    final uint8List =  await addLogo(path, id, suffix: "/logo");
+    return Image.memory(uint8List);
   }
 }
